+++ src/main/java/de/samply/frontend/Action.java
@@ -3,6 +3,7 @@ package de.samply.frontend;
 public record Action(
         String path,
         String method,
-        String[] params
+        String[] params,
+        String explanation
 ) {
 }